add moveImage functionality
